chatdba/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.20
chromadb==0.4.15
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
cors==1.0.1
# DeepSeek 支持
deepseek-sdk==0.1.0